+++ core/src/index/knowhere/knowhere/index/vector_index/FaissBaseIndex.cpp
@@ -33,7 +33,7 @@ FaissBaseIndex::SerializeImpl() {
     try {
         faiss::Index* index = index_.get();
 
-        SealImpl();
+        // SealImpl();
 
         MemoryIOWriter writer;
         faiss::write_index(index, &writer);
@@ -60,6 +60,8 @@ FaissBaseIndex::LoadImpl(const BinarySet& index_binary) {
     faiss::Index* index = faiss::read_index(&reader);
 
     index_.reset(index);
+
+    SealImpl();
 }
 
 void